Move ConstChildren into MRConstChildren.h and document how to use it
