+++ source/MRMesh/MRObject.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "MRPch/MRBindingMacros.h"
 #include "MRAffineXf3.h"
+#include "MRConstChildren.h"
 #include "MRExpected.h"
 #include "MRProgressCallback.h"
 #include "MRSignal.h"
@@ -10,7 +10,6 @@
 #include <array>
 #include <filesystem>
 #include <future>
-#include <iterator>
 #include <memory>
 #include <set>
 #include <vector>
@@ -28,44 +27,6 @@ namespace MR
  * \brief This chapter represents documentation about data models
  * \{
  */
-
-/// read-only access to the children of an Object, see Object::constChildren()
-class ConstChildren
-{
-public:
-    using Storage = std::vector<std::shared_ptr<Object>>;
-
-    class MR_BIND_IGNORE_PY Iterator
-    {
-    public:
-        using iterator_category = std::input_iterator_tag;
-        using value_type = std::shared_ptr<const Object>;
-        using difference_type = std::ptrdiff_t;
-
-        Iterator() = default;
-        explicit Iterator( Storage::const_iterator it ) : it_( it ) {}
-
-        [[nodiscard]] value_type operator *() const { return *it_; }
-        Iterator & operator ++() { ++it_; return *this; }
-        Iterator operator ++( int ) { auto res = *this; ++it_; return res; }
-        [[nodiscard]] friend bool operator ==( const Iterator & a, const Iterator & b ) { return a.it_ == b.it_; }
-
-    private:
-        Storage::const_iterator it_;
-    };
-
-    explicit ConstChildren( const Storage & children ) : children_( children ) {}
-
-    [[nodiscard]] MR_BIND_IGNORE_PY Iterator begin() const { return Iterator( children_.begin() ); }
-    [[nodiscard]] MR_BIND_IGNORE_PY Iterator end() const { return Iterator( children_.end() ); }
-
-    [[nodiscard]] bool empty() const { return children_.empty(); }
-    [[nodiscard]] size_t size() const { return children_.size(); }
-    [[nodiscard]] std::shared_ptr<const Object> operator []( size_t i ) const { return children_[i]; }
-
-private:
-    const Storage & children_;
-};
 
 /// the main purpose of this class is to avoid copy and move constructor and assignment operator
 /// implementation in Object class, which has too many fields for that;
